# docs/openapi3.0.yml
openapi: 3.0.3
info:
  title: SimpleQ - API documentation
  version: 1.0.0
  description: The API specification for the SimpleQ project
servers:
  - url: http://localhost:3000
components:
  schemas:
    Author: 
      type: object
      properties:
        id:   
          type: string
          description: the id of the author
        name: 
          type: string
        type: 
          type: string
          enum: ["guest", "user", "pro", "ai"]
    UserContent:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        likes:
          type: integer
        dislikes:
          type: integer
        created:
          type: string
          description: The date, the question was created
        opinion:
          type: string
          enum:
            - like
            - dislike
            - none
    Question:
      allOf:
        - $ref: '#/components/schemas/UserContent'
        - type: object
          properties:
            title:
              type: string
            tags: 
              type: array
              items:
                type: string
            numberOfAnswers:
              type: integer
            updated: 
              type: string
              description: The date, the question was last updated
            author: 
                $ref: '#/components/schemas/Author'
    Answer:
      allOf:
        - $ref: '#/components/schemas/UserContent'
        - type: object
          properties:
            author:
              $ref:  '#/components/schemas/Author'
...
    CreateQuestion:
      type: object
      properties:
        title:
          type: string
          nullable: false
        tags:
          type: array
          items:
            type: string
        content: 
          type: string
          nullable: false
        useAI:
          type: boolean
          description: Will be validated in the backend for user permission
          default: false
    CreateAnswer:
      type: object
      properties:
        content:
          type: string
          nullable: false
...
paths:
...
  /question/my:
    get:
      tags:
        - questions
      parameters:
        - in: query
          name: sortBy
          required: false
          description: Sort the answer by given criteria
          schema:
            type: string
            enum: [ ldr, likes, dislikes, timestamp ]
            default: ldr
        - in: query
          name: sortDirection
          required: false
          description: The direction of order, either ASC or DESC
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
        - in: query
          name: offset
          required: false
          description: Ignore x amount of answers
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          required: false
          description: Limit the amount of answers
          schema:
            type: integer
            default: 10
            minimum: 0
            maximum: 50
      summary: Get your own questions
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '401':
          description: 'Unauthorized, user needs to be logged in'
  /question/search/:
    get:
      tags:
        - questions
      summary: 'Get a list of question for specific search request, empty list if none'
      parameters:
        - in: query
          name: q # searchString
          description: The string to search for
          required: true
          schema: 
            type: string
        - in: query
          name: sortBy
          required: false
          description: Sort the answer by given criteria
          schema: 
            type: string
            enum: [ldr, likes, dislikes, timestamp]
            default: ldr
        - in: query
          name: sortDirection
          required: false
          description: The direction of order, either ASC or DESC
          schema: 
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: offset
          required: false
          description: Ignore x amount of answers
          schema: 
            type: integer
            default: 0
        - in: query
          name: limit
          required: false
          description: Limit the amount of answers
          schema: 
            type: integer
            default: 10
            minimum: 0
            maximum: 50
      responses:
        '200':
          description: 'Successful; **TODO: maybe skip content in this view to lower response payload**'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
  /question/{questionId}:
    get: 
      tags:
        - questions
      summary: Get a details to a single question
      parameters:
        - in: path
          name: questionId
          required: true
          description: UUID of a question
          schema:
            type: string
      responses:
        '200':
          description: Successfully fetched a single question
          content:
            application/json: 
              schema:
                allOf:
                  - $ref: '#/components/schemas/Question'
                  - type: object
                    properties:
                      isFavourite:
                        type: boolean
        '404':
          description: Not found
    delete:
      summary: Delete the content of a single question
      tags:
        - questions
      parameters:
        - in: path
          name: questionId
          required: true
          description: UUID of a question
          schema:
            type: string
      responses:
        '200':
          description: Successfully deleted the content of this question
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
        '404':
          description: No question found with this id.
        '401':
          description: Unauthorized, user must be logged in
  /question/{questionId}/title:
    get: 
      tags:
        - questions
      summary: Get a title for one question
      parameters:
        - in: path
          name: questionId
          required: true
          description: UUID of a question
          schema:
            type: string
      responses:
        '200':
          description: Successfully fetched a single title
          content:
            application/json: 
              schema: 
                properties:
                  title:
                    type: string
        '404':
          description: Not found
  /question/{questionId}/answers:
    get: 
      tags:
        - questions
      summary: Get a answers to a single question
      parameters:
        - in: path
          name: questionId
          required: true
          description: UUID of a question
          schema:
            type: string
        - in: query
          name: sortBy
          required: false
          description: Sort the answer by given criteria
          schema: 
            type: string
            enum: [ldr, likes, dislikes, timestamp]
            default: ldr
        - in: query
          name: sortDirection
          required: false
          description: The direction of order, either ASC or DESC
          schema: 
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: offset
          required: false
          description: Ignore x amount of answers
          schema: 
            type: integer
            default: 0
        - in: query
          name: limit
          required: false
          description: Limit the amount of answers
          schema: 
            type: integer
            default: 10
            minimum: 0
            maximum: 50
        - in: query
          name: enableAI
          required: false
          description: Should the AI answer be returned
          schema: 
            type: boolean
            default: true
      responses:
        '200':
          description: Successfully fetched a single question
          content:
            application/json: 
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Answer'
        '404':
          description: Not found      
  /question/create:
    post: 
      tags: 
       - questions
      summary: Create a new question
      requestBody:
        description: the data that needs to be send
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestion'
      responses:
        '200': 
          description: 'Question successfully created'
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    userId:
                      type: string
        '422':
          description: 'Unprocessable Entity: Payload is not sufficient'
  /question/{questionId}/answer:
    post:
      tags:
       - questions
      summary: Create a answer for a question
      parameters:
        - in: path
          name: questionId
          required: true
          description: UUID of a question
          schema:
            type: string
      requestBody:
        description: the data that needs to be send
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnswer'
      responses:
        '200':
          description: 'Answer successfully created'
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Answer'
        '422':
          description: 'Unprocessable Entity: Payload is not sufficient'
        '404':
          description: 'Not Found'
        '401':
          description: 'Forbidden, user needs to be registered'
        '406':
          description: 'Not Acceptable, used forbidden words in request'
...